SVG plot: fix issue with PHP 8.1
FossilOrigin-Name: 609e89ee27a1d76fa095053d260461ee43ca5669

+++ src/lib/KD2/Graphics/SVG/Plot.php
@@ -179,7 +179,7 @@ class Plot
 		$column_width = 70 + $this->data[0]->width;
 		$nb_items = ceil($axis_width / $column_width);
 		$item_width = $axis_width / $this->count;
-		$step = max(1, $this->count / $nb_items);
+		$step = (int) max(1, $this->count / $nb_items);
 
 		$i = 0;
